+++ tablet_qt/tasklib/inittasks.cpp
@@ -52,6 +52,7 @@
 #include "tasks/contactlog.h"
 #include "tasks/copebrief.h"
 #include "tasks/core10.h"
+#include "tasks/cpftcovidmedical.h"
 #include "tasks/cpftlpsdischarge.h"
 #include "tasks/cpftlpsreferral.h"
 #include "tasks/cpftlpsresetresponseclock.h"
@@ -206,6 +207,7 @@ void InitTasks(TaskFactory& factory)
     initializeContactLog(factory);
     initializeCopeBrief(factory);
     initializeCore10(factory);
+    initializeCPFTCovidMedical(factory);
     initializeCPFTLPSDischarge(factory);
     initializeCPFTLPSReferral(factory);
     initializeCPFTLPSResetResponseClock(factory);