+++ docs/source/changelog.rst
@@ -3462,3 +3462,4 @@ Current C++/SQLite client, Python/SQLAlchemy server
 - Fix Debian package Python dependencies. Would fail if Python 3.6 was not installed.
 
 - **New task:** :ref:`CPFT Research Preferences  <cpft_research_preferences>` (Database revision 0067).
+- **New task:** :ref:`CPFT Post-Covid Clinic Medical Questionnaire  <cpft_covid_medical>` (Database revision 0068).